Create main
this is a little bit of a complicated macro for editing the format of client billing reports from Practice Management to a certain easy to use format and then creating a pivot table from them to save managers time and effort. It's a WIP, not there yet. Yes some of this code is extraneous