This branch avoids using .Select Method for optim
optimizing code. It seems to work. I want to preserve the original as that worked and I don't want to break it or determine where back I had to go to. So two branches, the first works with .select and the second without .select - and any more optimizations I may come up with.