added my entire macro collection from my add-in
In case I need to access later and to secure it. 

One macro on there was copied from GitHub, called by me "Exporter". When set up right, it works to export all the macros of a workbook to files in a directory, but alas I couldn't figure out how to make it export my add-in modules. Perhaps I will figure that out someday. 

I think it has to do with line 25. VBProject is a property of a workbook object per the docs. Does an .xlam constitute such a workbook object? There is an add-in object per the docs but they don't contain a vbproject property. They contain much fewer properties. Just thought I'd share.